# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Meken API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/auth/anonymous": {
      "post": {
        "tags": [
          "identity"
        ],
        "summary": "Anonymous",
        "operationId": "anonymous_v1_auth_anonymous_post",
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "session_expired: missing, invalid or expired session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "apartment_not_found, conversation_not_found or turn_not_found. Resources owned by another user also return not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "idempotency_conflict, turn_in_progress, favorites_limit, conversations_limit or conversation_limit. Reusing an idempotency key with a different request is a conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "request_too_large: request body exceeds the limit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "invalid_request: body or path/query validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "rate_limited: wait for Retry-After seconds before retrying.",
            "headers": {
              "Retry-After": {
                "description": "Delay in seconds.",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "temporarily_unavailable: retry with backoff.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/me": {
      "get": {
        "tags": [
          "identity"
        ],
        "summary": "Profile",
        "operationId": "profile_v1_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "401": {
            "description": "session_expired: missing, invalid or expired session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "apartment_not_found, conversation_not_found or turn_not_found. Resources owned by another user also return not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "idempotency_conflict, turn_in_progress, favorites_limit, conversations_limit or conversation_limit. Reusing an idempotency key with a different request is a conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "request_too_large: request body exceeds the limit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "invalid_request: body or path/query validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "rate_limited: wait for Retry-After seconds before retrying.",
            "headers": {
              "Retry-After": {
                "description": "Delay in seconds.",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "temporarily_unavailable: retry with backoff.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "identity"
        ],
        "summary": "Delete Profile",
        "operationId": "delete_profile_v1_me_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "session_expired: missing, invalid or expired session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "apartment_not_found, conversation_not_found or turn_not_found. Resources owned by another user also return not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "idempotency_conflict, turn_in_progress, favorites_limit, conversations_limit or conversation_limit. Reusing an idempotency key with a different request is a conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "request_too_large: request body exceeds the limit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "invalid_request: body or path/query validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "rate_limited: wait for Retry-After seconds before retrying.",
            "headers": {
              "Retry-After": {
                "description": "Delay in seconds.",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "temporarily_unavailable: retry with backoff.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/v1/config": {
      "get": {
        "tags": [
          "discovery"
        ],
        "summary": "Config",
        "operationId": "config_v1_config_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppConfiguration"
                }
              }
            }
          },
          "401": {
            "description": "session_expired: missing, invalid or expired session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "apartment_not_found, conversation_not_found or turn_not_found. Resources owned by another user also return not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "idempotency_conflict, turn_in_progress, favorites_limit, conversations_limit or conversation_limit. Reusing an idempotency key with a different request is a conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "request_too_large: request body exceeds the limit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "invalid_request: body or path/query validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "rate_limited: wait for Retry-After seconds before retrying.",
            "headers": {
              "Retry-After": {
                "description": "Delay in seconds.",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "temporarily_unavailable: retry with backoff.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/apartments/{apartment_id}": {
      "get": {
        "tags": [
          "discovery"
        ],
        "summary": "Detail",
        "operationId": "detail_v1_apartments__apartment_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "apartment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Apartment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Listing"
                }
              }
            }
          },
          "401": {
            "description": "session_expired: missing, invalid or expired session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "apartment_not_found, conversation_not_found or turn_not_found. Resources owned by another user also return not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "idempotency_conflict, turn_in_progress, favorites_limit, conversations_limit or conversation_limit. Reusing an idempotency key with a different request is a conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "request_too_large: request body exceeds the limit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "invalid_request: body or path/query validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "rate_limited: wait for Retry-After seconds before retrying.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Delay in seconds."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "temporarily_unavailable: retry with backoff.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/apartments/{apartment_id}/verify": {
      "post": {
        "tags": [
          "discovery"
        ],
        "summary": "Verify",
        "operationId": "verify_v1_apartments__apartment_id__verify_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "apartment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Apartment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationResponse"
                }
              }
            }
          },
          "401": {
            "description": "session_expired: missing, invalid or expired session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "apartment_not_found, conversation_not_found or turn_not_found. Resources owned by another user also return not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "idempotency_conflict, turn_in_progress, favorites_limit, conversations_limit or conversation_limit. Reusing an idempotency key with a different request is a conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "request_too_large: request body exceeds the limit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "invalid_request: body or path/query validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "rate_limited: wait for Retry-After seconds before retrying.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Delay in seconds."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "temporarily_unavailable: retry with backoff.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/favorites": {
      "get": {
        "tags": [
          "discovery"
        ],
        "summary": "Favorites",
        "operationId": "favorites_v1_favorites_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingPage"
                }
              }
            }
          },
          "401": {
            "description": "session_expired: missing, invalid or expired session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "apartment_not_found, conversation_not_found or turn_not_found. Resources owned by another user also return not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "idempotency_conflict, turn_in_progress, favorites_limit, conversations_limit or conversation_limit. Reusing an idempotency key with a different request is a conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "request_too_large: request body exceeds the limit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "invalid_request: body or path/query validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "rate_limited: wait for Retry-After seconds before retrying.",
            "headers": {
              "Retry-After": {
                "description": "Delay in seconds.",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "temporarily_unavailable: retry with backoff.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/v1/conversations/{conversation_id}/turns": {
      "post": {
        "tags": [
          "discovery"
        ],
        "summary": "Send Turn",
        "operationId": "send_turn_v1_conversations__conversation_id__turns_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Conversation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TurnRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Durable event stream. A replay is a snapshot ending with pending while the turn runs, or done when it has terminated. Persist accepted.turn_id (also supplied as X-Turn-ID when sending a turn) and the last durable event id to resume safely.",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string",
                  "$ref": "#/components/schemas/SSEStreamSchema"
                }
              }
            }
          },
          "401": {
            "description": "session_expired: missing, invalid or expired session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "apartment_not_found, conversation_not_found or turn_not_found. Resources owned by another user also return not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "idempotency_conflict, turn_in_progress, favorites_limit, conversations_limit or conversation_limit. Reusing an idempotency key with a different request is a conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "request_too_large: request body exceeds the limit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "invalid_request: body or path/query validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "rate_limited: wait for Retry-After seconds before retrying.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Delay in seconds."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "temporarily_unavailable: retry with backoff.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "APIErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/APIError"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "APIErrorResponse"
      },
...
      "AppConfiguration": {
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "demo",
              "live"
            ],
            "title": "Mode"
          },
          "ai_enabled": {
            "type": "boolean",
            "title": "Ai Enabled"
          },
          "cities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Cities"
          },
          "privacy_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Privacy Url"
          },
          "terms_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Terms Url"
          },
          "retention_days": {
            "type": "integer",
            "title": "Retention Days"
          },
          "capabilities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Capabilities",
            "description": "Supported optional API features; absent means legacy."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mode",
          "ai_enabled",
          "cities",
          "privacy_url",
          "terms_url",
          "retention_days"
        ],
        "title": "AppConfiguration"
      },
...
      "ListingPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Listing"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ListingPage"
      },
...
      "ProfileResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "identity": {
            "type": "string",
            "const": "anonymous",
            "title": "Identity"
          },
          "retention_days": {
            "type": "integer",
            "title": "Retention Days"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "identity",
          "retention_days"
        ],
        "title": "ProfileResponse"
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "access_token",
          "refresh_token",
          "expires_in",
          "user_id"
        ],
        "title": "TokenResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}